refactor: 카드 뭉치 생성 책임 이전
